１回目レビュー後修正

+++ app/controllers/items_controller.rb
@@ -1,7 +1,7 @@
 class ItemsController < ApplicationController
 
-  before_action :authenticate_user!, only: [:new, :edit]
-  before_action :set_item, only: [:show, :edit, :update]
+  before_action :authenticate_user!, only: [:new, :edit, :destroy]
+  before_action :set_item, only: [:show, :edit, :update, :destroy]
 
 
   def index
@@ -39,7 +39,11 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    @item = Item.find(params[:id])
+    if current_user != @item.user
+      redirect_to item_path(@item.id)
+      return
+    end
+    
     if @item.destroy
       redirect_to root_path
     end
